requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
httpx==0.27.0
python-multipart==0.0.20
pillow
python-dotenv==1.0.1
psycopg2-binary==2.9.10
supabase==2.10.0

